Cast magma_int_t values to (int) in printf, to avoid warnings when 64-bit. Delete some unused variables. (in testing)
git-svn-id: https://icl.cs.utk.edu/svn/magma/trunk@3078 02bb5f2a-90ca-11de-bb24-8aedcc835bc8

+++ testing/testing_dlarfg.cpp
@@ -94,7 +94,7 @@ int main( int argc, char** argv)
             blasf77_daxpy( &size, &c_neg_one, h_x1, &ione, h_x2, &ione);
             error = lapackf77_dlange( "F", &N, &nb, h_x2, &N, work );
             printf("%5d %5d   %7.2f (%7.2f)   %7.2f (%7.2f)   %8.2g\n",
-                   (int) N, nb, cpu_perf, 1000.*cpu_time, gpu_perf, 1000.*gpu_time, error );
+                   (int) N, (int) nb, cpu_perf, 1000.*cpu_time, gpu_perf, 1000.*gpu_time, error );
             
             TESTING_FREE( h_x   );
             TESTING_FREE( h_x1  );
