+++ testing/testing_zgetrf_mgpu.cpp
@@ -85,7 +85,7 @@ int main( int argc, char** argv )
     magma_opts opts;
     parse_opts( argc, argv, &opts );
 
-    printf("ngpu %d\n", opts.ngpu );
+    printf("ngpu %d\n", (int) opts.ngpu );
     printf("    M     N   CPU GFlop/s (sec)   GPU GFlop/s (sec)   ||PA-LU||/(||A||*N)\n");
     printf("=========================================================================\n");
     for( int i = 0; i < opts.ntest; ++i ) {
