+++ testing/testing_zpotrf.cpp
@@ -40,7 +40,7 @@ int main( int argc, char** argv)
     parse_opts( argc, argv, &opts );
     opts.lapack |= opts.check;  // check (-c) implies lapack (-l)
     
-    printf("ngpu %d, uplo %c\n", opts.ngpu, opts.uplo );
+    printf("ngpu %d, uplo %c\n", (int) opts.ngpu, opts.uplo );
     printf("    N   CPU GFlop/s (sec)   GPU GFlop/s (sec)   ||R_magma - R_lapack||_F / ||R_lapack||_F\n");
     printf("========================================================\n");
     for( int i = 0; i < opts.ntest; ++i ) {
