+++ testing/testing_zunmqr.cpp
@@ -121,7 +121,7 @@ int main( int argc, char** argv )
                        (int) info, magma_strerror( info ));
             lwork = (magma_int_t) MAGMA_Z_REAL( W[0] );
             if ( lwork < 0 || lwork > lwork_max )
-                printf("invalid lwork %d, lwork_max %d\n", lwork, lwork_max );
+                printf("invalid lwork %d, lwork_max %d\n", (int) lwork, (int) lwork_max );
             
             gpu_time = magma_wtime();
             magma_zunmqr( *side[iside], *trans[itran],
