+++ testing/testing_zunmqr_gpu.cpp
@@ -136,7 +136,7 @@ int main( int argc, char** argv )
                        (int) info, magma_strerror( info ));
             lwork = (magma_int_t) MAGMA_Z_REAL( W[0] );
             if ( lwork < 0 || lwork > lwork_max )
-                printf("invalid lwork %d, lwork_max %d\n", lwork, lwork_max );
+                printf("invalid lwork %d, lwork_max %d\n", (int) lwork, (int) lwork_max );
             
             gpu_time = magma_sync_wtime( 0 );  // sync needed for L,N and R,T cases
             magma_zunmqr_gpu( *side[iside], *trans[itran],
